Also restructure the packages for the mqtt integration tests. (#5667)
