+++ itests/org.openhab.binding.mqtt.homeassistant.tests/pom.xml
@@ -19,6 +19,11 @@
       <artifactId>org.openhab.binding.mqtt</artifactId>
       <version>${project.version}</version>
     </dependency>
+      <dependency>
+          <groupId>org.openhab.addons.bundles</groupId>
+          <artifactId>org.openhab.binding.mqtt.generic</artifactId>
+          <version>${project.version}</version>
+      </dependency>
     <dependency>
       <groupId>org.openhab.addons.bundles</groupId>
       <artifactId>org.openhab.binding.mqtt.homeassistant</artifactId>
