+++ itests/org.openhab.binding.mqtt.homeassistant.tests/src/main/java/org/openhab/binding/mqtt/DiscoverComponentsTests.java
@@ -28,13 +28,12 @@ import org.eclipse.smarthome.test.java.JavaOSGiTest;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
-import org.openhab.binding.mqtt.TransformationServiceProvider;
-import org.openhab.binding.mqtt.internal.handler.ThingChannelConstants;
-import org.openhab.binding.mqtt.internal.homeassistant.ChannelConfigurationTypeAdapterFactory;
-import org.openhab.binding.mqtt.internal.homeassistant.DiscoverComponents;
-import org.openhab.binding.mqtt.internal.homeassistant.HaID;
-import org.openhab.binding.mqtt.internal.homeassistant.HandlerConfiguration;
-import org.openhab.binding.mqtt.internal.homeassistant.DiscoverComponents.ComponentDiscovered;
+import org.openhab.binding.mqtt.generic.TransformationServiceProvider;
+import org.openhab.binding.mqtt.homeassistant.internal.ChannelConfigurationTypeAdapterFactory;
+import org.openhab.binding.mqtt.homeassistant.internal.DiscoverComponents;
+import org.openhab.binding.mqtt.homeassistant.internal.HaID;
+import org.openhab.binding.mqtt.homeassistant.internal.HandlerConfiguration;
+import org.openhab.binding.mqtt.homeassistant.internal.DiscoverComponents.ComponentDiscovered;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
