+++ itests/org.openhab.binding.mqtt.homie.tests/src/main/java/org/openhab/binding/mqtt/generic/HomieImplementationTests.java
@@ -44,20 +44,19 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 import org.mockito.invocation.InvocationOnMock;
-import org.openhab.binding.mqtt.ChannelState;
-import org.openhab.binding.mqtt.internal.handler.HomieThingHandler;
-import org.openhab.binding.mqtt.internal.homie300.Device;
-import org.openhab.binding.mqtt.internal.homie300.DeviceAttributes;
-import org.openhab.binding.mqtt.internal.homie300.DeviceAttributes.ReadyState;
-import org.openhab.binding.mqtt.internal.homie300.DeviceCallback;
-import org.openhab.binding.mqtt.internal.homie300.Node;
-import org.openhab.binding.mqtt.internal.homie300.NodeAttributes;
-import org.openhab.binding.mqtt.internal.homie300.Property;
-import org.openhab.binding.mqtt.internal.homie300.PropertyAttributes;
-import org.openhab.binding.mqtt.internal.homie300.PropertyAttributes.DataTypeEnum;
-import org.openhab.binding.mqtt.internal.homie300.PropertyHelper;
-import org.openhab.binding.mqtt.tools.ChildMap;
-import org.openhab.binding.mqtt.tools.WaitForTopicValue;
+import org.openhab.binding.mqtt.homie.internal.handler.HomieThingHandler;
+import org.openhab.binding.mqtt.homie.internal.homie300.Device;
+import org.openhab.binding.mqtt.homie.internal.homie300.DeviceAttributes;
+import org.openhab.binding.mqtt.homie.internal.homie300.DeviceAttributes.ReadyState;
+import org.openhab.binding.mqtt.homie.internal.homie300.DeviceCallback;
+import org.openhab.binding.mqtt.homie.internal.homie300.Node;
+import org.openhab.binding.mqtt.homie.internal.homie300.NodeAttributes;
+import org.openhab.binding.mqtt.homie.internal.homie300.Property;
+import org.openhab.binding.mqtt.homie.internal.homie300.PropertyAttributes;
+import org.openhab.binding.mqtt.homie.internal.homie300.PropertyAttributes.DataTypeEnum;
+import org.openhab.binding.mqtt.homie.internal.homie300.PropertyHelper;
+import org.openhab.binding.mqtt.generic.tools.ChildMap;
+import org.openhab.binding.mqtt.generic.tools.WaitForTopicValue;
 import org.osgi.service.cm.ConfigurationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
