+++ itests/org.openhab.binding.mqtt.homie.tests/src/main/java/org/openhab/binding/mqtt/homie/internal/homie300/PropertyHelper.java
@@ -10,11 +10,11 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.mqtt.internal.homie300;
+package org.openhab.binding.mqtt.homie.internal.homie300;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.mqtt.ChannelState;
+import org.openhab.binding.mqtt.generic.ChannelState;
 
 /**
  * Helper to access {@link Property} internals.
